+++ forgerock-auth/src/test/java/org/forgerock/android/auth/AuthTestSuite.java
@@ -24,7 +24,7 @@ import org.junit.runners.Suite;
         TextOutputCallbackTest.class,
         ReCaptchaCallbackTest.class,
         ConsentMappingCallbackTest.class,
-        DeviceAttributeCollectorCallbackTest.class,
+        DeviceProfileCollectorCallbackTest.class,
 
         AuthServiceMockTest.class,
         AuthServiceTest.class,